% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsr.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map fatal accidents on a state polygon for a given year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State code (integer)}

\item{year}{Year (integer)}
}
\value{
A map of fatal accidents for a given state and year.
}
\description{
\code{fars_maps_state} generates a maps of fatal accidents for
a given state and year. An error is raised for invalid state numbers.
A warning is displayed if a year is not present in the data set.
}
\examples{
\donttest{
fars_map_state(state.num = 11, year = 2013)
fars_map_state(17, 2014)
}

}
