% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsr.R
\name{fars_read}
\alias{fars_read}
\title{Read the FARS (Fatality Analysis Reporting System) data set}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A single string, path to a file, or a connection.}
}
\value{
The FARS data set (tibble data frame).
}
\description{
\code{fars_read} reads the FARS data set from disk. This is a wrapper around the
\code{readr} function from the \code{readr} package. An error will be raised if the
file does not exist.
}
\examples{
\donttest{
fars_read(filename = './raw-data/accident_2013.csv.bz2')
}

}
