% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsr.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read the FARS data from disk for a sequence of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of numbers (years).}
}
\value{
A list of data frames or NULL if the file cannot be found
}
\description{
\code{fars_read_years} reads the FARS data from disk for input years and outputs
a list of data frames. A warning will be generated if data for a given year are
not available.
}
\examples{
\donttest{
fars_read_years(2015)
}

}
